refactor: refactor into folders for maintainability
