+++ src/nextTick.js
@@ -1,7 +1,17 @@
 'use strict'
 
+let promise = Promise.resolve()
 let mutateWithTask
 let currTask
+
+module.exports = function nextTick (task) {
+  currTask = task
+  if (mutateWithTask) {
+    mutateWithTask()
+  } else {
+    promise = promise.then(task)
+  }
+}
 
 if (typeof MutationObserver !== 'undefined') {
   let counter = 0
@@ -9,23 +19,14 @@ if (typeof MutationObserver !== 'undefined') {
   const textNode = document.createTextNode(String(counter))
   observer.observe(textNode, {characterData: true})
 
-  function onTask () {
-    if (currTask) {
-      currTask()
-    }
-  }
-
   mutateWithTask = function mutateWithTask () {
     counter = (counter + 1) % 2
     textNode.textContent = counter
   }
 }
 
-module.exports = function nextTick (task) {
-  currTask = task
-  if (mutateWithTask) {
-    mutateWithTask()
-  } else {
-    Promise.resolve().then(task)
+function onTask () {
+  if (currTask) {
+    currTask()
   }
 }
